Update playbooks for GCP
Preparing to change hosting to Google Compute Platform, the
playbooks are updated to use a modern version of Ansible by
using remote_user and become* keys. The become is necessary
as GCP appropriately configures to enable a non-root user for
ssh connections only.

Inventory is changed to point to a temporary host IP in
preparation for changing DNS entries.

A new playbook is introduced to ensure the host is hardened
for some of the obvious host configuration options