Revise deploy process
Revised site deploy to speed up execution and streamline processing.
It's still a but of an R.G. device but is immensely faster and more
elegant.

Since the output directory is in .gitignore it can become it's own
sandbox area WRT git. It is now expected to be configured as a git
repo of it's own (remote shows the same repo) on the 'deploy'
branch.

The 'deploy' branch only contains content to be published. That
branch can be checked out directly into the web root.

nikola deploy directive is configured to use scripts/deploy_site.sh
which will automatically commit and push a new version of the deploy
branch to github and trigger ansible playbook push-content.yml to
the live site from the deploy branch on github.